Disable tab hydrate on App start. Move Context Menu to post_cell. Fix issue with triggering Edit navigation to page & post when long pressing for ContextMenu on Android. Remove "Edit" option if page is a template.
